requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudinary==1.33.0
cryptography==41.0.3
Deprecated==1.2.14
Django==4.2.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dlib==19.24.2
drf-yasg==1.21.6
face-recognition==1.3.0
face-recognition-models==0.3.0
frozenlist==1.4.0
googlemaps==4.10.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
multidict==6.0.4
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.25.0
packaging==23.1
Pillow==9.5.0
protobuf==3.20.3
pycparser==2.21
pydantic==1.10.12
PyJWT==2.7.0
PyMySQL==1.0.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
twilio==8.5.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
vonage==3.7.1
vonage-jwt==1.0.0
wrapt==1.15.0
yarl==1.9.2
